@@ -23,26 +23,28 @@
     "rc": "^1.2.8",
     "redis": "^3.0.2",
     "winston": "^3.0.0",
-    "ws": "^7.4.2"
+    "ws": "^7.4.4",
+    "pg": "^8.5.1"
   },
   "devDependencies": {
     "@types/body-parser": "^1.17.1",
     "@types/jest": "^26.0.20",
-    "@types/node": "^14.14.20",
-    "@types/node-forge": "^0.8.4",
-    "@typescript-eslint/eslint-plugin": "^4.9.1",
-    "@typescript-eslint/parser": "^4.9.1",
-    "eslint": "^7.15.0",
+    "@types/node": "^14.14.33",
+    "@types/node-forge": "^0.9.7",
+    "@types/pg": "^7.14.11",
+    "@typescript-eslint/eslint-plugin": "^4.17.0",
+    "@typescript-eslint/parser": "^4.17.0",
+    "eslint": "^7.21.0",
     "eslint-config-standard": "^16.0.2",
-    "eslint-config-standard-with-typescript": "^19.0.1",
+    "eslint-config-standard-with-typescript": "^20.0.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-promise": "^4.3.1",
     "eslint-plugin-standard": "^5.0.0",
     "jest": "26.6.3",
     "jest-transform-stub": "^2.0.0",
-    "ts-jest": "^26.4.4",
-    "ts-node": "^8.1.0",
-    "typescript": "^3.6.2"
+    "ts-jest": "^26.5.3",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.3"
   }
 }
