@@ -6,19 +6,20 @@
 import * as fs from 'fs'
 import * as path from 'path'
 import { logger as log } from './utils/logger'
-import { mpsMicroservice } from './mpsMicroservice'
+import { MPSMicroservice } from './mpsMicroservice'
 import { configType, certificatesType } from './models/Config'
-import { dataBase } from './utils/db'
+import { Database } from './utils/db'
 
 import { certificates } from './utils/certificates'
 import { tlsConfig } from './utils/tlsConfiguration'
 import { IDbProvider } from './models/IDbProvider'
 
 import { SecretManagerService } from './utils/SecretManagerService'
-import { secretsDbProvider } from './utils/vaultDbProvider'
+import { SecretsDbProvider } from './utils/vaultDbProvider'
 import { parseValue } from './utils/parseEnvValue'
 
-const rc = require('rc')
+import rc from 'rc'
+import { Environment } from './utils/Environment'
 
 try {
 // To merge ENV variables. consider after lowercasing ENV since our config keys are lowercase
@@ -39,17 +40,17 @@ try {
   // path where Self-signed certificates are generated
   const certPath = path.join(__dirname, config.cert_path)
   config.data_path = path.join(__dirname, config.data_path)
-
   let certs: certificatesType
   let db: IDbProvider
   log.silly(`Updated config... ${JSON.stringify(config, null, 2)}`)
+  Environment.Config = config
 
   // DB initialization
   if (config.use_vault) {
     log.info('Using secrets db provider')
-    db = new secretsDbProvider(new SecretManagerService(config, log), log, config)
+    db = new SecretsDbProvider(new SecretManagerService(config, log), log, config)
   } else {
-    db = new dataBase(config)
+    db = new Database(config)
   }
   // Certificate Configuration and Operations
   if (config.https || !config.tls_offload) {
@@ -80,7 +81,9 @@ try {
       }
       log.info('Loaded existing certificates')
     } else {
-      if (!fs.existsSync(certPath)) { fs.mkdirSync(certPath, { recursive: true }) }
+      if (!fs.existsSync(certPath)) {
+        fs.mkdirSync(certPath, { recursive: true })
+      }
       certs = certificates.generateCertificates(config, certPath)
     }
 
@@ -89,10 +92,10 @@ try {
     // TODO: comment this out for release
      //log.silly(`certs: ` + JSON.stringify(certs,null, 4));
 
-    const mps = new mpsMicroservice(config, db, certs)
+    const mps = new MPSMicroservice(config, db, certs)
     mps.start()
   }
 } catch (error) {
-  log.error('Error starting mps Microservice. Check server logs.')
+  log.error('Error starting MPS microservice. Check server logs.')
   log.error(error)
 }
