@@ -7,19 +7,19 @@
 
 import { ISecretManagerService } from '../models/ISecretManagerService'
 import { configType } from '../models/Config'
-import NodeVault = require('node-vault');
+import NodeVault = require('node-vault')
 
 export class SecretManagerService implements ISecretManagerService {
-  vaultClient: NodeVault.client;
-  logger: any;
-  constructor (config:configType, logger: any, vault?: any) {
+  vaultClient: NodeVault.client
+  logger: any
+  constructor (config: configType, logger: any, vault?: any) {
     this.logger = logger
     if (vault) {
       this.vaultClient = vault
       return
     }
 
-    const options : NodeVault.VaultOptions = {
+    const options: NodeVault.VaultOptions = {
       apiVersion: 'v1', // default
       endpoint: config.vault_address, // default
       token: config.vault_token // optional client token; can be fetched after valid initialization of the server
@@ -28,7 +28,7 @@ export class SecretManagerService implements ISecretManagerService {
     this.vaultClient = NodeVault(options)
   }
 
-  async listSecretsAtPath (path:string): Promise<any> {
+  async listSecretsAtPath (path: string): Promise<any> {
     try {
       this.logger.info('list secret ' + path)
       const data = await this.vaultClient.list(path)
@@ -43,7 +43,7 @@ export class SecretManagerService implements ISecretManagerService {
     }
   }
 
-  async getSecretFromKey (path:string, key: string): Promise<string> {
+  async getSecretFromKey (path: string, key: string): Promise<string> {
     try {
       this.logger.info('getting secret ' + path + ' ' + key)
       const data = await this.vaultClient.read(path)
@@ -55,19 +55,19 @@ export class SecretManagerService implements ISecretManagerService {
       this.logger.error(error)
       return null
     }
-  }  
+  }
 
-  async getSecretAtPath(path:string): Promise<any> {
+  async getSecretAtPath (path: string): Promise<any> {
     try {
       this.logger.info('getting secrets from ' + path)
-      let data = await this.vaultClient.read(path);
+      const data = await this.vaultClient.read(path)
       // this.logger.info(`got data back from vault : ${data}`)
-      return data.data; 
+      return data.data
     } catch (error) {
       this.logger.error('getSecretAtPath error \r\n')
       this.logger.error(error)
-      return null;
-    }  
+      return null
+    }
   }
 
   async readJsonFromKey (path: string, key: string): Promise<string> {
@@ -75,7 +75,7 @@ export class SecretManagerService implements ISecretManagerService {
     return (data ? JSON.parse(data) : null)
   }
 
-  async writeSecretWithKey (path: string, key:string, keyValue: any): Promise<void> {
+  async writeSecretWithKey (path: string, key: string, keyValue: any): Promise<void> {
     const data = { data: {} }
     data.data[key] = keyValue
     // this.logger.info('writing:' + JSON.stringify(data))
