@@ -1,4 +1,4 @@
-import { mpsMicroservice } from "./mpsMicroservice";
+import { MPSMicroservice } from "./mpsMicroservice";
 export class CiraChannelFactory {
   factoryFn: (socket: any, port: any) => any
   constructor( factoryFn: (socket:any, port:any) => any) {
