@@ -7,22 +7,23 @@
 import { Response, Request } from 'express'
 import { logger as log } from '../../utils/logger'
 import { IAmtHandler } from '../../models/IAmtHandler'
-import { mpsMicroservice } from '../../mpsMicroservice'
-import { amtStackFactory, amtPort } from '../../utils/constants'
+import { MPSMicroservice } from '../../mpsMicroservice'
+import { amtPort } from '../../utils/constants'
 import { ErrorResponse } from '../../utils/amtHelper'
+import AMTStackFactory from '../../amt_libraries/amt-connection-factory.js'
 
 export class PowerStateHandler implements IAmtHandler {
-  mpsService: mpsMicroservice;
-  name: string;
-  amtFactory: any;
+  mpsService: MPSMicroservice
+  name: string
+  amtFactory: any
 
-  constructor (mpsService: mpsMicroservice) {
+  constructor (mpsService: MPSMicroservice) {
     this.name = 'PowerState'
     this.mpsService = mpsService
-    this.amtFactory = new amtStackFactory(this.mpsService)
+    this.amtFactory = new AMTStackFactory(this.mpsService)
   }
 
-  async AmtAction (req: Request, res: Response) {
+  async AmtAction (req: Request, res: Response): Promise<void> {
     try {
       const payload = req.body.payload
       if (payload.guid) {
@@ -35,21 +36,22 @@ export class PowerStateHandler implements IAmtHandler {
             if (status != 200) {
               log.error(`Request failed during powerstate fetch for guid : ${payload.guid}.`)
               return res.status(status).send(ErrorResponse(status, `Request failed during powerstate fetch for guid : ${payload.guid}.`))
+            } else {
+              const resbody = { powerstate: responses[0].PowerState }
+              res.send(resbody)
             }
-            const resbody = { powerstate: responses[0].PowerState }
-            return res.send(resbody)
           })
         } else {
           res.set({ 'Content-Type': 'application/json' })
-          return res.status(404).send(ErrorResponse(404, `guid : ${payload.guid}`, 'device'))
+          res.status(404).send(ErrorResponse(404, `guid : ${payload.guid}`, 'device'))
         }
       } else {
         res.set({ 'Content-Type': 'application/json' })
-        return res.status(404).send(ErrorResponse(404, null, 'guid'))
+        res.status(404).send(ErrorResponse(404, null, 'guid'))
       }
     } catch (error) {
       log.error(`Exception in Power state : ${error}`)
-      return res.status(500).send(ErrorResponse(500, 'Request failed during powerstate fetch.'))
+      res.status(500).send(ErrorResponse(500, 'Request failed during powerstate fetch.'))
     }
   }
 }
