@@ -8,25 +8,26 @@
 import { Response, Request } from 'express'
 import { logger as log } from '../../utils/logger'
 import { IAmtHandler } from '../../models/IAmtHandler'
-import { mpsMicroservice } from '../../mpsMicroservice'
+import { MPSMicroservice } from '../../mpsMicroservice'
 
-import { amtStackFactory, amtPort, UserConsentOptions } from '../../utils/constants'
+import { amtPort, UserConsentOptions } from '../../utils/constants'
 import { ErrorResponse } from '../../utils/amtHelper'
 import { AMTFeatures } from '../../utils/AMTFeatures'
+import AMTStackFactory from '../../amt_libraries/amt-connection-factory.js'
 
 import { MPSValidationError } from '../../utils/MPSValidationError'
 import { apiResponseType } from '../../models/Config'
 
 export class SetAMTFeaturesHandler implements IAmtHandler {
-    mpsService: mpsMicroservice;
-    amtFactory: any;
-    name: string;
+  mpsService: MPSMicroservice
+  amtFactory: any
+  name: string
 
-    constructor (mpsService: mpsMicroservice) {
-      this.name = 'SetAMTFeatures'
-      this.mpsService = mpsService
-      this.amtFactory = new amtStackFactory(this.mpsService)
-    }
+  constructor (mpsService: MPSMicroservice) {
+    this.name = 'SetAMTFeatures'
+    this.mpsService = mpsService
+    this.amtFactory = new AMTStackFactory(this.mpsService)
+  }
 
     async AmtAction (req: Request, res: Response) {
       try {
@@ -46,45 +47,43 @@ export class SetAMTFeaturesHandler implements IAmtHandler {
             res.set({ 'Content-Type': 'application/json' })
             return res.status(404).send(ErrorResponse(404, `guid : ${payload.guid}`, 'device'))
           }
-        } else {
-          res.set({ 'Content-Type': 'application/json' })
-          return res.status(404).send(ErrorResponse(404, null, 'guid'))
         }
-      } catch (error) {
-        log.error(`Exception in set AMT Features: ${error}`)
-        if (error instanceof MPSValidationError) {
-          return res.status(error.status || 400).send(ErrorResponse(error.status || 400, error.message))
-        }
-        return res.status(500).send(ErrorResponse(500, 'Request failed during set AMT Features.'))
+    } catch (error) {
+      log.error(`Exception in set AMT Features: ${error}`)
+      if (error instanceof MPSValidationError) {
+        res.status(error.status || 400).send(ErrorResponse(error.status || 400, error.message))
+      } else {
+        res.status(500).send(ErrorResponse(500, 'Request failed during set AMT Features.'))
       }
     }
+  }
 
-    validatePayload (payload: any) {
-      if (payload.userConsent === undefined && payload.enableSOL === undefined &&
+  validatePayload (payload: any): void {
+    if (payload.userConsent === undefined && payload.enableSOL === undefined &&
             payload.enableIDER === undefined && payload.enableKVM === undefined) {
-        throw new MPSValidationError(`Device : ${payload.guid} to set AMT features,at least on flag is mandatory.\n userConsent:"kvm/all/none" \n enableRedir: true/false \n enableSOL: true/false \n enableIDER: true/false \n enableKVM: true/false`)
-      }
+      throw new MPSValidationError(`Device : ${payload.guid} to set AMT features,at least on flag is mandatory.\n userConsent:"kvm/all/none" \n enableRedir: true/false \n enableSOL: true/false \n enableIDER: true/false \n enableKVM: true/false`)
+    }
 
-      // Check if valid option is received. 0:None,1:For KVM only, 0xFFFFFFFF: All
-      if (payload.userConsent !== undefined) {
-        if (typeof payload.userConsent === 'string') {
-          const key = payload.userConsent.toLowerCase()
-          if (!UserConsentOptions.hasOwnProperty(key)) {
-            throw new MPSValidationError(`Device : ${payload.guid} User Consent should be "kvm/all/none"`)
-          }
-        } else {
+    // Check if valid option is received. 0:None,1:For KVM only, 0xFFFFFFFF: All
+    if (payload.userConsent !== undefined) {
+      if (typeof payload.userConsent === 'string') {
+        const key = payload.userConsent.toLowerCase()
+        if (!UserConsentOptions.hasOwnProperty(key)) {
           throw new MPSValidationError(`Device : ${payload.guid} User Consent should be "kvm/all/none"`)
         }
-      }
-
-      if (payload.enableSOL !== undefined && typeof payload.enableSOL !== 'boolean') {
-        throw new MPSValidationError(`Device : ${payload.guid} enableSOL should be boolean`)
-      }
-      if (payload.enableIDER !== undefined && typeof payload.enableIDER !== 'boolean') {
-        throw new MPSValidationError(`Device : ${payload.guid} enableIDER should be boolean`)
-      }
-      if (payload.enableKVM !== undefined && typeof payload.enableKVM !== 'boolean') {
-        throw new MPSValidationError(`Device : ${payload.guid} enableKVM should be boolean`)
+      } else {
+        throw new MPSValidationError(`Device : ${payload.guid} User Consent should be "kvm/all/none"`)
       }
     }
+
+    if (payload.enableSOL !== undefined && typeof payload.enableSOL !== 'boolean') {
+      throw new MPSValidationError(`Device : ${payload.guid} enableSOL should be boolean`)
+    }
+    if (payload.enableIDER !== undefined && typeof payload.enableIDER !== 'boolean') {
+      throw new MPSValidationError(`Device : ${payload.guid} enableIDER should be boolean`)
+    }
+    if (payload.enableKVM !== undefined && typeof payload.enableKVM !== 'boolean') {
+      throw new MPSValidationError(`Device : ${payload.guid} enableKVM should be boolean`)
+    }
+  }
 }
