@@ -8,21 +8,21 @@ import { logger as log } from '../../utils/logger'
 import { IAdminHandler } from '../../models/IAdminHandler'
 import { Response, Request } from 'express'
 import { ErrorResponse } from '../../utils/amtHelper'
-import { mpsMicroservice } from '../../mpsMicroservice'
+import { MPSMicroservice } from '../../mpsMicroservice'
 import { MpsProxy } from '../../server/proxies/MpsProxy'
 
 export class DisconnectHandler implements IAdminHandler {
-  mps: mpsMicroservice;
-  name: string;
+  mps: MPSMicroservice
+  name: string
 
-  constructor (mps: mpsMicroservice) {
+  constructor (mps: MPSMicroservice) {
     this.name = 'Disconnect'
     this.mps = mps
   }
 
   // Get existing device list from credentials file.
   // For the server version of Mesh Commander, we send the computer list without credential and insertion credentials in the stream.
-  async adminAction (req: Request, res: Response) {
+  async adminAction (req: Request, res: Response): Promise<void> {
     try {
       const payload = req.body.payload
       // Check if request body contains guid information
@@ -32,28 +32,20 @@ export class DisconnectHandler implements IAdminHandler {
         if (ciraconn) {
           try {
             ciraconn.destroy()
-            res.set({ 'Content-Type': 'application/json' })
-            return res.send(
-              JSON.stringify(
-                `{ success: 200, description: 'CIRA connection disconnected : ${payload.guid}'}`
-              )
-            )
+            res.json({ success: 200, description: `CIRA connection disconnected : ${payload.guid}` })
           } catch (error) {
             log.error(error)
-            res.set({ 'Content-Type': 'application/json' })
-            return res.status(500).send(ErrorResponse(500, error))
+            res.status(500).json(ErrorResponse(500, error))
           }
         } else {
-          res.set({ 'Content-Type': 'application/json' })
-          return res.status(404).send(ErrorResponse(404, `guid : ${payload.guid}`, 'device'))
+          res.status(404).json(ErrorResponse(404, `guid : ${payload.guid}`, 'device'))
         }
       } else {
-        res.set({ 'Content-Type': 'application/json' })
-        return res.status(404).send(ErrorResponse(404, null, 'guid'))
+        res.status(404).json(ErrorResponse(404, null, 'guid'))
       }
     } catch (error) {
       log.error(`Exception in Disconnect: ${JSON.stringify(error)} `)
-      return res.status(500).send(ErrorResponse(500, 'Request failed while disconnecting device.'))
+      res.status(500).json(ErrorResponse(500, 'Request failed while disconnecting device.'))
     }
   }
 }
