@@ -4,22 +4,22 @@
 * Description: Routes all the admin endpoint post calls
 **********************************************************************/
 import { AdminController } from '../controllers/adminController'
-import { mpsMicroservice } from '../mpsMicroservice'
+import { MPSMicroservice } from '../mpsMicroservice'
 import * as express from 'express'
 
-export class adminRoutes {
-  mpsService: mpsMicroservice;
-  adminController: AdminController;
-  router: any;
+export class AdminRoutes {
+  mpsService: MPSMicroservice
+  adminController: AdminController
+  router: any
 
-  constructor (mpsService: mpsMicroservice) {
+  constructor (mpsService: MPSMicroservice) {
     this.mpsService = mpsService
     AdminController.init(this.mpsService)
     this.router = express.Router()
     this.buildRoutes()
   }
 
-  buildRoutes () : any {
+  buildRoutes (): any {
     this.router.post('/', AdminController.HandlePostRoute)
   }
 }
