@@ -10,9 +10,9 @@ import { AMTFeaturesConst, UserConsentOptions } from './constants'
 
 export class AMTFeatures {
   static async getAMTFeatures (amtstack: any, payload: any): Promise<any> {
-    return new Promise((resolve, reject) => {
+    return await new Promise((resolve, reject) => {
       amtstack.BatchEnum('', ['*IPS_OptInService', '*AMT_RedirectionService', '*CIM_KVMRedirectionSAP'], (stack, name, wsmanResponse, status) => {
-        if (status != 200) {
+        if (status !== 200) {
           reject(new MPSValidationError(`Failed AMT features BatchEnum Exec for guid : ${payload.guid}.`))
         } else {
           resolve(wsmanResponse)
@@ -21,7 +21,7 @@ export class AMTFeatures {
     })
   }
 
-  static async setAMTFeatures (amtstack: any, payload: any) {
+  static async setAMTFeatures (amtstack: any, payload: any): Promise<void> {
     const wsmanResponse = await AMTFeatures.getAMTFeatures(amtstack, payload)
     const amtRedirResponse = wsmanResponse[AMTFeaturesConst.AMT_REDIR_SERVICE].response
     const kvmRedirResponse = wsmanResponse[AMTFeaturesConst.AMT_KVM_REDIR].response
@@ -30,9 +30,9 @@ export class AMTFeatures {
     if (wsmanResponse[AMTFeaturesConst.AMT_REDIR_SERVICE] && wsmanResponse[AMTFeaturesConst.AMT_KVM_REDIR]) {
       let redir, sol, ider, kvm
 
-      redir = (amtRedirResponse[AMTFeaturesConst.AMT_REDIR_LISTENER] == true)
-      sol = ((amtRedirResponse[AMTFeaturesConst.AMT_REDIR_STATE] & 2) != 0)
-      ider = ((amtRedirResponse[AMTFeaturesConst.AMT_REDIR_STATE] & 1) != 0)
+      redir = (amtRedirResponse[AMTFeaturesConst.AMT_REDIR_LISTENER] === true)
+      sol = ((amtRedirResponse[AMTFeaturesConst.AMT_REDIR_STATE] & 2) !== 0)
+      ider = ((amtRedirResponse[AMTFeaturesConst.AMT_REDIR_STATE] & 1) !== 0)
 
       if (payload.enableSOL !== undefined && payload.enableSOL !== sol) {
         sol = payload.enableSOL
@@ -46,8 +46,8 @@ export class AMTFeatures {
         isRedirectionChanged = true
       }
 
-      kvm = ((kvmRedirResponse[AMTFeaturesConst.AMT_KVM_ENABLED_STATE] == 6 && kvmRedirResponse[AMTFeaturesConst.AMT_KVM_REQUESTED_STATE] == 2) ||
-                kvmRedirResponse[AMTFeaturesConst.AMT_KVM_ENABLED_STATE] == 2 || kvmRedirResponse[AMTFeaturesConst.AMT_KVM_ENABLED_STATE] == 6)
+      kvm = ((kvmRedirResponse[AMTFeaturesConst.AMT_KVM_ENABLED_STATE] === 6 && kvmRedirResponse[AMTFeaturesConst.AMT_KVM_REQUESTED_STATE] === 2) ||
+                kvmRedirResponse[AMTFeaturesConst.AMT_KVM_ENABLED_STATE] === 2 || kvmRedirResponse[AMTFeaturesConst.AMT_KVM_ENABLED_STATE] === 6)
       if (payload.enableKVM !== undefined && payload.enableKVM !== kvm) {
         kvm = payload.enableKVM
         isRedirectionChanged = true
@@ -77,18 +77,18 @@ export class AMTFeatures {
     }
   }
 
-  static async setRedirectionService (amtstack: any, amtRedirResponse: any, kvm: any, payload: any) {
-    return new Promise((resolve, reject) => {
+  static async setRedirectionService (amtstack: any, amtRedirResponse: any, kvm: any, payload: any): Promise<boolean> {
+    return await new Promise((resolve, reject) => {
       amtstack.AMT_RedirectionService_RequestStateChange(amtRedirResponse.EnabledState, (stack, name, response, status) => {
-        if (status != 200) {
+        if (status !== 200) {
           reject(new MPSValidationError(`Failed Redir Request State Change for guid : ${payload.guid}.`, status))
         } else {
           amtstack.CIM_KVMRedirectionSAP_RequestStateChange(kvm ? 2 : 3, 0, (stack, name, response, status) => {
-            if (status != 200) {
+            if (status !== 200) {
               reject(new MPSValidationError(`Failed to update KVM Redirection Service for guid : ${payload.guid}.`, status))
             }
-            amtstack.Put(AMTFeaturesConst.AMT_REDIR_SERVICE, amtRedirResponse, (stack, name, kvm_redir_response, status) => {
-              if (status != 200) {
+            amtstack.Put(AMTFeaturesConst.AMT_REDIR_SERVICE, amtRedirResponse, (stack, name, kvmRedirResponse, status) => {
+              if (status !== 200) {
                 reject(new MPSValidationError(`Failed AMT RedirectionService update for guid : ${payload.guid}.`, status))
               }
               resolve(true)
@@ -99,10 +99,10 @@ export class AMTFeatures {
     })
   }
 
-  static async setUserConsent (amtstack: any, optServiceRes: any, payload: any) {
-    return new Promise((resolve, reject) => {
+  static async setUserConsent (amtstack: any, optServiceRes: any, payload: any): Promise<boolean> {
+    return await new Promise((resolve, reject) => {
       amtstack.Put(AMTFeaturesConst.AMT_OPTIN_SERVICE, optServiceRes, (stack, name, response, status) => {
-        if (status != 200) {
+        if (status !== 200) {
           reject(new MPSValidationError(`Failed to get OptInService guid : ${payload.guid}.`, status))
         }
         resolve(true)
