@@ -9,8 +9,8 @@ import * as fs from 'fs'
 
 import { logger as log } from './logger'
 import { mpsConfigType, webConfigType, directConfigType } from '../models/Config'
+import { constants } from 'crypto'
 
-const constants = require('crypto').constants ? require('crypto').constants : require('constants')
 const webTlsConfigPath = path.join(__dirname, '../../private/webtlsconfig.json')
 const mpsTlsConfigPath = path.join(__dirname, '../../private/mpstlsconfig.json')
 const directConnTlsConfigPath = path.join(__dirname, '../../private/directConntlsconfig.json')
@@ -36,7 +36,7 @@ export class tlsConfig {
           delete webConfig[i]
         }
         if (webConfig[i] instanceof Array) {
-          if (webConfig[i].length == 0) {
+          if (webConfig[i].length === 0) {
             delete webConfig[i]
             continue
           }
@@ -49,7 +49,7 @@ export class tlsConfig {
           !fs.existsSync(path.join(__dirname, webConfig.key)) &&
           !fs.existsSync(path.join(__dirname, webConfig.cert))
         ) {
-          log.error('Error: TLS cerficate or private key does not exist.')
+          log.error('Error: TLS certificate or private key does not exist.')
           process.exit()
         } else {
           webConfig.key = fs.readFileSync(
@@ -91,13 +91,12 @@ export class tlsConfig {
       // Perform 'OR' operation between SecureOptions
       // Example: { secureOptions: constants.SSL_OP_NO_SSLv2 | constants.SSL_OP_NO_SSLv3 |  constants.SSL_OP_NO_TLSv1 | constants.SSL_OP_NO_TLSv11}
       if (webConfig.secureOptions) {
-        if (webConfig.secureOptions.length == 1) {
+        if (webConfig.secureOptions.length === 1) {
           // No need of 'OR' if only one option
-          webConfig.secureOptions = constants.webConfig.secureOptions[0]
+          webConfig.secureOptions = webConfig.secureOptions[0]
         } else {
           const optionArr = webConfig.secureOptions
-          let secoption: any =
-            constants[optionArr[0]] | constants[optionArr[1]]
+          let secoption: any = constants[optionArr[0]] | constants[optionArr[1]]
           for (let i: number = 2; i < optionArr.length; i++) {
             secoption = secoption | constants[optionArr[i]]
           }
@@ -134,7 +133,7 @@ export class tlsConfig {
           continue
         }
         if (mpsConfig[i] instanceof Array) {
-          if (mpsConfig[i].length == 0) {
+          if (mpsConfig[i].length === 0) {
             delete mpsConfig[i]
             continue
           }
@@ -169,9 +168,9 @@ export class tlsConfig {
       // Perform 'OR' operation between SecureOptions
       // Example: { secureOptions: constants.SSL_OP_NO_SSLv2 | constants.SSL_OP_NO_SSLv3 |  constants.SSL_OP_NO_TLSv1 | constants.SSL_OP_NO_TLSv11}
       if (mpsConfig.secureOptions) {
-        if (mpsConfig.secureOptions.length == 1) {
+        if (mpsConfig.secureOptions.length === 1) {
           // No need of 'OR' if only one option
-          mpsConfig.secureOptions = constants.mpsConfig.secureOptions[0]
+          mpsConfig.secureOptions = mpsConfig.secureOptions[0]
         } else {
           const optionArr = mpsConfig.secureOptions
           let secoption: any =
