@@ -5,12 +5,13 @@
 import { IDbProvider } from '../models/IDbProvider'
 import { ISecretManagerService } from '../models/ISecretManagerService'
 import { configType } from '../models/Config'
+import { Credentials } from '../models/models'
 
-export class secretsDbProvider implements IDbProvider {
+export class SecretsDbProvider implements IDbProvider {
   secretsManager: ISecretManagerService
-  secretsPath: string;
-  logger: any;
-  config: configType;
+  secretsPath: string
+  logger: any
+  config: configType
   constructor (secretsManager: ISecretManagerService, logger: any, config: configType) {
     this.secretsManager = secretsManager
     this.secretsPath = config.secrets_path
@@ -18,7 +19,7 @@ export class secretsDbProvider implements IDbProvider {
     this.config = config
   }
 
-  async CIRAAuth (guid: string, username: string, password: string, cb: any) {
+  async CIRAAuth (guid: string, username: string, password: string, cb: any): Promise<boolean> {
     try {
       // let user = await this.secretsManager.getSecretFromKey(`${this.secretsPath}devices/${guid}`, `username`);
       // let pwd = await this.secretsManager.getSecretFromKey(`${this.secretsPath}devices/${guid}`, `password`);
@@ -32,7 +33,9 @@ export class secretsDbProvider implements IDbProvider {
         pwd = this.config.pass
 
         if (username === user && password === pwd) {
-          if (cb) { cb(true) }
+          if (cb) {
+            cb(true)
+          }
           return
         }
         this.logger.info('invalid mps credentials')
@@ -42,7 +45,9 @@ export class secretsDbProvider implements IDbProvider {
         pwd = await this.secretsManager.getSecretFromKey(`${this.secretsPath}devices/${guid}`, 'mps_password')
 
         if (username === user && password === pwd) {
-          if (cb) { cb(true) }
+          if (cb) {
+            cb(true)
+          }
           return
         }
         this.logger.info('invalid mps credentials')
@@ -58,8 +63,8 @@ export class secretsDbProvider implements IDbProvider {
   async getAmtPassword (guid: string): Promise<string[]> {
     try {
       const user = 'admin'
-      let data: any = await this.secretsManager.getSecretAtPath(`${this.secretsPath}devices/${guid}`)  
-      let amtpass = data.data['AMT_PASSWORD']
+      const data: any = await this.secretsManager.getSecretAtPath(`${this.secretsPath}devices/${guid}`)
+      const amtpass = data.data.AMT_PASSWORD
       return [user, amtpass]
     } catch (error) {
       this.logger.error('Error while retrieving device credentials\r\n')
@@ -68,11 +73,11 @@ export class secretsDbProvider implements IDbProvider {
     }
   }
 
-  async IsGUIDApproved (guid: string, cb: any) {
+  async IsGUIDApproved (guid: string, cb: any): Promise<void> {
     try {
       let result = false
       const guids = await this.secretsManager.getSecretFromKey(`${this.secretsPath}global`, 'guids_allowlist')
-      if (guids.indexOf(guid) >= 0) {
+      if (guids.includes(guid)) {
         result = true
       }
       if (cb) {
@@ -84,12 +89,12 @@ export class secretsDbProvider implements IDbProvider {
     }
   }
 
-  async IsOrgApproved (org: string, cb: any) {
+  async IsOrgApproved (org: string, cb: any): Promise<void> {
     try {
       let result = false
       const orgs = await this.secretsManager.getSecretFromKey(`${this.secretsPath}global`, 'orgs_allowlist')
 
-      if (orgs.indexOf(org) >= 0) {
+      if (orgs.includes(org)) {
         result = true
       }
       if (cb) {
@@ -104,10 +109,9 @@ export class secretsDbProvider implements IDbProvider {
     }
   }
 
-  async getAllAmtCredentials (): Promise<any> {
+  async getAllAmtCredentials (): Promise<Credentials> {
     try {
       try {
-        const result = false
         const path = this.secretsPath.replace('data', 'metadata')
         const creds = await this.secretsManager.listSecretsAtPath(`${path}devices`)
 
